fix: adding snapshot support
