+++ packages/c-accordion/tests/c-accordion.cy.tsx
