+++ packages/c-icon/tests/c-icon.cy.tsx
