+++ packages/tests/env.ts
@@ -1,0 +1,2 @@
+window.global = window;
+window.process = window.process || { env: {}, argv: [] }
