+++ packages/tests/snapshots.ts
@@ -1,0 +1,2 @@
+import * as cySnapshots from '@cypress/snapshot'
+cySnapshots.register()
