+++ packages/tests/styles.ts
@@ -1,0 +1,36 @@
+const style = document.createElement('style')
+
+// TODO: Import a shared css file that playground/App.vue uses
+// Should be as simple as `import '../../playground/App.vue?style`
+style.innerText = `html, body {
+  margin: 0;
+  height: 100vh;
+  width: 100vw;
+}
+
+html {
+  line-height: 1.5;
+  color: rgb(26, 32, 44);
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+}
+
+#app {
+  height: inherit;
+  width: inherit;
+}
+
+a {
+  text-decoration: none;
+}
+
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.3s ease-in-out;
+}
+
+.fade-enter-from,
+.fade-leave-to {
+  opacity: 0;
+}`
+
+document.head.appendChild(style)
