+++ snapshots.js
@@ -1,0 +1,10 @@
+module.exports = {
+  __version: '6.9.0',
+  '<Portal />': {
+    'Portal Examples': {
+      'renders SimplePortal successfully': {
+        1: '<div data-v-app="">\n  <div class="css-16m79av">\n    <section class="css-0"> Open Vue Dev Tools to inspect page for portal\n      elements. <span class="css-0"></span></section>\n    <main class="css-0">\n      <!--teleport start-->\n      <!--teleport end-->\n    </main>\n  </div>\n  <!---->\n</div>',
+      },
+    },
+  },
+}
